+++ src/modules/products/components/ProductList.tsx
@@ -6,19 +6,19 @@ export const ProductList: React.FC = () => {
   const [page, setPage] = useState(1);
   const { data, isLoading, isError } = useGetProducts({page});
   const products = data?.content ?? [];
-  const totalElements = data?.totalElements ?? 0;
-  const pageSize = 12; 
-  const totalPages = Math.ceil(totalElements / pageSize);
-
-
-
+  
   if (isLoading) return <div>Loading...</div>;
   if (isError) return <div>ERROR</div>
 
   return (
-    <div className="container mx-auto p-10">
+    <div>
       <h1>Product List</h1>
-      <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-5">
+      <div className="gap-5 grid justify-center place-content-center items-center auto-cols-fr"
+      style={{
+        gridTemplateColumns: "repeat(auto-fit, minmax(200px, 1fr))",
+        gridAutoRows: "22rem"
+      }}
+      >
         {products.map((product) => (
           <ProductCard key={product.productId} product={product} />
         ))}
@@ -37,7 +37,7 @@ export const ProductList: React.FC = () => {
         <button
           className="px-4 py-2 bg-gray-300 rounded"
           onClick={() => setPage(Math.max(page + 1, 1))}
-          disabled={page >= totalPages}>
+          disabled={page === 1}>
           Next →
         </button>
       </div>
