--- src/modules/products/interfaces/product.interface.ts
@@ -1,9 +1,0 @@
-export interface Product {
-    productId: string;
-    name: string;
-    description: string;
-    price: number;
-    stock: number;
-    color: string;
-    imageURLS: string[];
-}
